+++ client/components/Dropdown.js
@@ -1,0 +1,12 @@
+import menuItems from './menuItems';
+export function Dropdown ({submenus}){
+  return (
+    <ul className="dropdown">
+      {submenus.map((submenu, index) => (
+        <li key={index} className="menu-items">
+          <a href={submenu.url}>{submenu.title}</a>
+        </li>
+      ))}
+    </ul>
+  );
+};
