+++ client/components/menuItems.js
@@ -1,19 +1,18 @@
-export const menuItems = [
- {
-   location: 'Pensacola',
-   url: '/',
- },
- {
-   location: 'Austin',
-   url: '/Austin',
- },
- {
-  location: 'London',
-  url: '/London',
- },
- {
-  location: 'New York',
-  url: '/NewYork',
- },
-];
+import {Dropdown} from './Dropdown' 
 
+export function MenuItems ({items}){
+  return (
+    <li className="menu-items">
+      {items.submenu ? (
+        <>
+          <button type="button" aria-haspopup="menu">
+            {items.title}{' '}
+          </button>
+          <Dropdown submenus={items.submenu} />
+        </>
+      ) : (
+        <a href={items.url}>{items.title}</a>
+      )}
+    </li>
+  );
+}
