+++ client/menuItems.js
@@ -1,0 +1,19 @@
+export const menuItems = [
+ {
+   location: 'Pensacola',
+   url: '/',
+ },
+ {
+   location: 'Austin',
+   url: '/Austin',
+ },
+ {
+  location: 'London',
+  url: '/London',
+ },
+ {
+  location: 'New York',
+  url: '/NewYork',
+ },
+];
+
